extract PaywallComponentsImagePreDownloader
